Removed the use of include comments to bring in chunks of code during building; All JS is contained within the main mousetip.js file
